1.1_Multi_Agent_Centralized_A2A/WriterAgent/requirements.txt
a2a-sdk[http-server]
uvicorn
httpx
langgraph
langchain
langchain-openai
langchain-community
pydantic
python-dotenv
tavily-python